safari-600.7-branch/LayoutTests/storage/domstorage/quota-expected.txt

Test the DOM Storage quota code.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing sessionStorage
storage.clear()
PASS storage.length is 0
Creating 'data' which contains 64K of data
PASS data.length is 65536
Putting 'data' into 39 sessionStorage buckets.
Putting 'data' into another bucket.h
PASS Insertion worked.


Testing localStorage
storage.clear()
PASS storage.length is 0
Creating 'data' which contains 64K of data
PASS data.length is 65536
Putting 'data' into 39 localStorage buckets.
Putting 'data' into another bucket.h
PASS Hit exception as expected
Verify that data was never inserted.
PASS storage.getItem(39) is null
Removing bucket 38.
Adding 'Hello!' into a new bucket.
PASS Insertion worked.
PASS successfullyParsed is true

TEST COMPLETE
